// src/interfaces/company.ts
export interface ICompany{
    id: number;
    uid: string;
    business_name: string;
    suffix?: string;
    industry?: string;
    catch_phrase?: string;
    buzzword?: string;
    bs_company_statement?: string;
    employee_identification_number?: string;
    duns_number?: string;
    logo?: string;
    type?: string;
    phone_number: string;
    full_address?: string;
    latitude?: number;
    longitude?: number;
}